+++ Data/Retail/ExportedCalendarEvents.lua
@@ -23,7 +23,7 @@ end
 -- [[ Everything after these lines is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2023-09-04 19-09-22 ]] --
+-- [[ Exported at 2023-09-30 11-18-51 ]] --
 tasks = {
     {N, 141, 236705}, -- Feast of Winter Veil
     {N, 181, 237000}, -- Noblegarden
@@ -39,12 +39,15 @@ tasks = {
     {N, 479, 134481}, -- Darkmoon Faire
     {N, 1262, 133783}, -- WoW's 18th Anniversary
     {N, 1395, 1100022}, -- Kalimdor Cup
-    {N, 1396, 4419345}, -- Secrets of Azeroth (US)
+    {N, 1396, 4419345}, -- Secrets of Azeroth
+    {N, 1398, 4419345}, -- Secrets of Azeroth
+    {N, 1399, 4419345}, -- Secrets of Azeroth
     {N, 1397, 133783}, -- WoW's 19th Anniversary
-    {N, 1398, 4419345}, -- Secrets of Azeroth (EU)
-    {N, 1399, 4419345}, -- Secrets of Azeroth (KR)
     {N, 1400, 1100022}, -- Eastern Kingdoms Cup
     {N, 1425, 4630413}, -- Turbulent Timeways
+    {N, 1458, 4630413}, -- Turbulent Timeways
+    {N, 1459, 4630413}, -- Turbulent Timeways
+    {N, 1460, 4630413}, -- Turbulent Timeways
 };
 
 function exportedCalendarEvents.LoadCategories(e)
@@ -63,34 +66,35 @@ function exportedCalendarEvents.LoadCategories(e)
     e[1262].Category = addon.Tabs["Events"].Categories[7]; -- WoW's Anniversary
     e[1395].Category = addon.Tabs["Events"].Categories[9].Children[1]; -- Kalimdor Cup
     e[1396].Category = addon.Tabs["Events"].Categories[10]; -- Secrets of Azeroth
-    e[1397].Category = addon.Tabs["Events"].Categories[7]; -- WoW's Anniversary
     e[1398].Category = addon.Tabs["Events"].Categories[10]; -- Secrets of Azeroth
     e[1399].Category = addon.Tabs["Events"].Categories[10]; -- Secrets of Azeroth
+    e[1397].Category = addon.Tabs["Events"].Categories[7]; -- WoW's Anniversary
     e[1400].Category = addon.Tabs["Events"].Categories[9].Children[2]; -- Eastern Kingdoms Cup
     e[1425].Category = addon.Tabs["Events"].Categories[11].Children[2]; -- Turbulent Timeways
+    e[1458].Category = addon.Tabs["Events"].Categories[11].Children[2]; -- Turbulent Timeways
+    e[1459].Category = addon.Tabs["Events"].Categories[11].Children[2]; -- Turbulent Timeways
+    e[1460].Category = addon.Tabs["Events"].Categories[11].Children[2]; -- Turbulent Timeways
 end
 
 function exportedCalendarEvents.InjectDynamicOptions()
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 141, addon.L["Feast of Winter Veil"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 181, addon.L["Noblegarden"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 201, addon.L["Children's Week"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 324, addon.L["Hallow's End"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 327, addon.L["Lunar Festival"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 341, addon.L["Midsummer Fire Festival"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 372, addon.L["Brewfest"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 398, addon.L["Pirates' Day"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 404, addon.L["Pilgrim's Bounty"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 409, addon.L["Day of the Dead"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], 423, addon.L["Love is in the Air"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {141}, addon.L["Feast of Winter Veil"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {181}, addon.L["Noblegarden"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {201}, addon.L["Children's Week"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {324}, addon.L["Hallow's End"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {327}, addon.L["Lunar Festival"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {341}, addon.L["Midsummer Fire Festival"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {372}, addon.L["Brewfest"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {398}, addon.L["Pirates' Day"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {404}, addon.L["Pilgrim's Bounty"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {409}, addon.L["Day of the Dead"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Holidays", addon.L["Holidays"], {423}, addon.L["Love is in the Air"]);
     KrowiAF_RegisterDeSelectAllEventOptions("Calendar", "Holidays", { 141, 181, 201, 324, 327, 341, 372, 398, 404, 409, 423 });
-    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], 479, addon.L["Darkmoon Faire"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], 1262, addon.L["WoW's 18th Anniversary"], false);
-    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], 1395, addon.L["Kalimdor Cup"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], 1396, addon.L["Secrets of Azeroth (US)"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], 1397, addon.L["WoW's 19th Anniversary"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], 1398, addon.L["Secrets of Azeroth (EU)"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], 1399, addon.L["Secrets of Azeroth (KR)"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], 1400, addon.L["Eastern Kingdoms Cup"]);
-    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], 1425, addon.L["Turbulent Timeways"]);
-    KrowiAF_RegisterDeSelectAllEventOptions("Calendar", "Other", { 479, 1262, 1395, 1396, 1397, 1398, 1399, 1400, 1425 });
+    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], {479}, addon.L["Darkmoon Faire"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], {1262}, addon.L["WoW's 18th Anniversary"], false);
+    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], {1395}, addon.L["Kalimdor Cup"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], {1396, 1398, 1399}, addon.L["Secrets of Azeroth"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], {1397}, addon.L["WoW's 19th Anniversary"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], {1400}, addon.L["Eastern Kingdoms Cup"]);
+    KrowiAF_RegisterEventOptions("Calendar", "Other", addon.L["Other"], {1425, 1458, 1459, 1460}, addon.L["Turbulent Timeways"]);
+    KrowiAF_RegisterDeSelectAllEventOptions("Calendar", "Other", { 479, 1262, 1395, 1396, 1398, 1399, 1397, 1400, 1425, 1458, 1459, 1460 });
 end